requirements.txt
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
asyncer==0.0.2
attrs==23.2.0
beautifulsoup4==4.12.3
bidict==0.23.1
blinker==1.8.2
bytesbufio==1.0.3
cachetools==5.3.3
caugetch==0.0.1
certifi==2024.2.2
cffi @ file:///E:/jobs/cffi-1.15.0-cp310-cp310-win_amd64.whl#sha256=bc34814757359d614add1392024fd7d09374b397d4222fa7c099b141c148ba2d
chainlit==1.0.505
charset-normalizer==3.3.2
chevron==0.14.0
click==8.1.7
clipboard==0.0.4
cohere==5.3.4
colorama==0.4.6
configparser==7.0.0
dataclasses-json==0.5.14
Deprecated==1.2.14
dnspython==2.6.1
exceptiongroup==1.2.1
faiss-cpu==1.8.0
fastapi==0.110.2
fastapi-socketio==0.0.10
fastavro==1.9.4
filelock==3.14.0
filetype==1.2.0
Flask==3.0.3
Flask-PyMongo==2.3.0
frozenlist==1.4.1
fsspec==2024.3.1
getpass4==0.0.14.1
google-ai-generativelanguage==0.6.2
google-api-core==2.18.0
google-api-python-client==2.127.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.2
googleapis-common-protos==1.63.0
googlesearch-python==1.2.4
greenlet==3.0.3
grpcio==1.62.2
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.22.2
idna==3.7
importlib-metadata==7.0.0
intel-openmp==2021.4.0
itsdangerous==2.2.0
Jinja2==3.1.3
joblib==1.4.0
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.16
langchain-community==0.0.34
langchain-core==0.1.52
langchain-google-genai==1.0.3
langchain-pinecone==0.1.1
langchain-text-splitters==0.0.1
langsmith==0.1.52
Lazify==0.4.0
literalai==0.0.507
MarkupSafe==2.1.5
marshmallow==3.21.1
mkl==2021.4.0
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-exporter-otlp-proto-http==1.24.0
opentelemetry-instrumentation==0.45b0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
orjson==3.10.1
packaging==23.2
pandas==2.2.2
pillow==10.3.0
pinecone-client==3.2.2
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.1
pydantic_core==2.18.2
PyJWT==2.8.0
pymongo==4.7.2
pyparsing==3.1.2
pypdf==4.2.0
PyPDF2==3.0.1
pyperclip==1.8.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.0
python-graphql-client==0.4.3
python-multipart==0.0.9
python-socketio==5.11.2
pytz==2024.1
PyYAML==6.0.1
regex==2024.4.28
requests==2.31.0
rsa==4.9
safetensors==0.4.3
scikit-learn==1.4.2
scipy==1.13.0
sentence-transformers==2.7.0
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.29
starlette==0.37.2
sympy==1.12
syncer==2.0.3
tbb==2021.12.0
tenacity==8.2.3
threadpoolctl==3.5.0
tokenizers==0.19.1
tomli==2.0.1
torch==2.3.0
tqdm==4.66.2
transformers==4.40.1
types-requests==2.31.0.20240406
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
uptrace==1.24.0
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.25.0
watchfiles==0.20.0
websockets==12.0
Werkzeug==3.0.3
wrapt==1.16.0
wsproto==1.2.0
xgboost==1.7.6
yarl==1.9.4
zipp==3.18.1